+++ packages/winnow/CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 1.16.0
 ### Added
 * Add normalization of a `option.sourceSR`; this option identifies the CRS of the source data and defaults to 'EPSG:4326'; 
 * If `option.sourceSR` is defined and geometry filter is defined, the geometry filter is reprojected to the CRS of the source
@@ -267,6 +267,7 @@ Deep clone for dates since mutating dates has down stream consequences if reused
 ### Added
 * Initial release
 
+[1.16.0]: https://github.com/featureserver/winnow/compare/v1.15.3...v1.16.0
 [1.15.3]: https://github.com/featureserver/winnow/compare/v1.15.2...v1.15.3
 [1.15.2]: https://github.com/featureserver/winnow/compare/v1.15.1...v1.15.2
 [1.15.1]: https://github.com/featureserver/winnow/compare/v1.15.0...v1.15.1